Move Google Analytics to a custom React hook
