+++ lib/apollo/withApollo.js
@@ -7,6 +7,7 @@ export default App => {
   return class Apollo extends React.Component {
     static displayName = 'withApollo(App)';
 
+    // TODO: This causes an error: https://err.sh/next.js/opt-out-automatic-prerendering
     static async getInitialProps (ctx) {
       const { Component, router } = ctx;
