+++ pages/_app.js
@@ -15,16 +15,6 @@ import '../style.css';
 const Auth = dynamic(() => import('../components/Auth'), { ssr: false });
 
 class MyApp extends App {
-  static async getInitialProps({ Component, ctx }) {
-    let pageProps = {};
-
-    if (Component.getInitialProps) {
-      pageProps = await Component.getInitialProps(ctx);
-    }
-
-    return { pageProps };
-  }
-
   componentDidCatch (error, errorInfo) {
     Sentry.withScope(scope => {
       Object.keys(errorInfo).forEach(key => {
